ENH: add integer-na support via an ExtensionArray
closes #20700
closes #20747

+++ pandas/core/series.py
@@ -4094,7 +4094,7 @@ def _sanitize_array(data, index, dtype=None, copy=False,
             elif is_extension_array_dtype(dtype):
                 # create an extension array from its dtype
                 array_type = dtype.construct_array_type()
-                subarr = array_type(subarr, copy=copy)
+                subarr = array_type(subarr, dtype=dtype, copy=copy)
 
             elif dtype is not None and raise_cast_failure:
                 raise
